Implements functionality to find the turning points within each ion chromatogram (to help separate isobars).
